MDF: rename json files to mdf
