+++ psyneulink/core/globals/__init__.py
@@ -1,6 +1,6 @@
 from . import context
 from . import defaults
-from . import json
+from . import mdf
 from . import keywords
 from . import kvo
 from . import log
@@ -12,10 +12,10 @@ from . import sampleiterator
 
 from .context import *
 from .defaults import *
-from .json import *
 from .keywords import *
 from .kvo import *
 from .log import *
+from .mdf import *
 from .parameters import *
 from .preferences import *
 from .registry import *
@@ -24,10 +24,10 @@ from .sampleiterator import *
 
 __all__ = list(context.__all__)
 __all__.extend(defaults.__all__)
-__all__.extend(json.__all__)
 __all__.extend(keywords.__all__)
 __all__.extend(kvo.__all__)
 __all__.extend(log.__all__)
+__all__.extend(mdf.__all__)
 __all__.extend(parameters.__all__)
 __all__.extend(preferences.__all__)
 __all__.extend(registry.__all__)
