+++ tests/mdf/model_with_two_disjoint_comps.py
